Refactor: Extract auth logic to auth.ts
`index.ts`에서 auth 기능을 `auth.ts`로 분리
