+++ packages/firebase/package.json
@@ -10,5 +10,8 @@
   "devDependencies": {
     "@workspace/typescript-config": "workspace:*",
     "typescript": "^5.9.2"
+  },
+  "exports": {
+    "./auth": "./src/auth.ts"
   }
 }
